shared hosting redirection issue fixed
In shared hosting, after activation(success or not) the redirection line was return $this->redirect('login'); which was directly calling "http://hostname.com/path/login" as like pretty urls are ON. But if pretty Urls are OFF, this redirection causes 500 error. Changing the redirection code to return "return $this->redirect(['login']);" fixed the issue.